requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-unfold==0.87.0
drf-spectacular==0.29.0
Pillow==11.2.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
